Moved all test fixtures into conftest.py [Finishes #133854409]
